- Moving 'fakes' into internal. The new namespace_test.go would cause us to have a circular reference otherwise. - Fixing up comment for the "use cached connection but still return 'true' to recover" code.
